+++ processor/src/test/resources/expected/com/example/repository/RepositoryWithExplicitIdRepository.java
@@ -34,7 +34,7 @@ public abstract class RepositoryWithExplicitIdRepository extends AbstractReposit
   }
 
   @Nullable
-  public RepositoryWithExplicitId findById(final int id) {
+  public final RepositoryWithExplicitId findById(final int id) {
     return findByArezId( id );
   }