+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 * **\[processor\]** The `create` methods on the generated repository incorrectly had their access level
   determined by the access level of the associated components constructor. This has been corrected so
   that the access level of the component class determines the access level of the method.
+* **\[processor\]** The methods on the generated repository and the repository class itself incorrectly
+  had defined the access level as public. This has been corrected so that the access level of the component
+  class determines the access level of the methods and the repository type.
 
 #### Changed
 * **\[core\]** Updated `ObserverErrorHandlerSupport` to improve dead code elimination in production mode.